+++ app/src/main/java/com/operatorsapp/managers/PersistenceManager.java
@@ -6,7 +6,6 @@ import android.content.Context;
 import com.app.operatorinfra.OperatorPersistenceManagerInterface;
 import com.example.oppapplog.OppAppLogger;
 import com.google.gson.Gson;
-import com.google.gson.JsonSyntaxException;
 import com.google.gson.reflect.TypeToken;
 import com.operators.activejobslistformachineinfra.ActiveJobsListForMachinePersistenceManagerInterface;
 import com.operators.infra.PersistenceManagerInterface;
@@ -96,7 +95,7 @@ public class PersistenceManager implements LoginPersistenceManagerInterface,
     private static PersistenceManager msInstance;
     private Gson mGson;
     @SuppressLint("UseSparseArrays")
-    public HashMap<Long, Event> items = new HashMap<>();
+    public HashMap<Float, Event> items = new HashMap<>();
 
     public static void initInstance(Context context) {
         if (msInstance == null) {
